requirements.txt
transformers==4.42.4
torch>=2.2.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
beautifulsoup4==4.12.3
trafilatura==1.8.0
requests==2.31.0
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
prefect==3.0.0
streamlit==1.36.0
tenacity==8.5.0
lxml_html_clean